refactor: single-round timers with rescheduling (#97)
## Summary

Refactors all timer-based processing functions to a single-round,
immediately-rescheduling pattern:

- Each timer invocation processes one round of up to
`MAX_CONCURRENT_RPC_CALLS` batches, then immediately reschedules itself
with `Duration::ZERO` if work remains
- Rescheduling uses a `scopeguard` that fires on drop (including
panics); defused when all work fits in the current round
- The `more_to_process` flag is computed from a snapshot of the queue
**before** processing begins, so items added concurrently during
execution don't trigger a spurious reschedule

## Stack

- Based on: #120 (split monitor timer)


---------

+++ minter/src/dashboard/tests.rs
@@ -207,8 +207,8 @@ fn should_paginate_minted_deposits_across_multiple_pages() {
     let remainder = total_deposits - DEFAULT_PAGE_SIZE * 2;
 
     for i in 0..total_deposits {
-        accept_deposit(deposit_id(i as u8), 500_000_000);
-        mint_deposit(deposit_id(i as u8), i as u64);
+        accept_deposit(deposit_id(i), 500_000_000);
+        mint_deposit(deposit_id(i), i as u64);
     }
 
     let page1 = dashboard();
